Solved my routing method probelm, and application now can access the correct routes using a new file in /lib.
