+++ app/controllers/application_controller.rb
@@ -1,3 +1,4 @@
+require 'media_routing.rb'
 class ApplicationController < ActionController::Base
   # Prevent CSRF attacks by raising an exception.
   # For APIs, you may want to use :null_session instead.
